+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/generators/ContributedInterfaceSupertypeGenerator.kt
@@ -17,7 +17,7 @@ import dev.zacsweers.metro.compiler.fir.resolvedExcludedClassIds
 import dev.zacsweers.metro.compiler.fir.resolvedReplacedClassIds
 import dev.zacsweers.metro.compiler.fir.resolvedScopeClassId
 import dev.zacsweers.metro.compiler.fir.scopeArgument
-import dev.zacsweers.metro.compiler.joinSimpleNames
+import dev.zacsweers.metro.compiler.flatMapToSet
 import dev.zacsweers.metro.compiler.singleOrError
 import java.util.TreeMap
 import org.jetbrains.kotlin.fir.FirAnnotationContainer
@@ -34,7 +34,9 @@ import org.jetbrains.kotlin.fir.expressions.FirArrayLiteral
 import org.jetbrains.kotlin.fir.extensions.FirDeclarationPredicateRegistrar
 import org.jetbrains.kotlin.fir.extensions.FirSupertypeGenerationExtension
 import org.jetbrains.kotlin.fir.extensions.predicateBasedProvider
+import org.jetbrains.kotlin.fir.lookupTracker
 import org.jetbrains.kotlin.fir.moduleData
+import org.jetbrains.kotlin.fir.recordFqNameLookup
 import org.jetbrains.kotlin.fir.resolve.providers.symbolProvider
 import org.jetbrains.kotlin.fir.resolve.toClassSymbol
 import org.jetbrains.kotlin.fir.resolve.toRegularClassSymbol
@@ -65,9 +67,8 @@ internal class ContributedInterfaceSupertypeGenerator(session: FirSession) :
   }
 
   private val inCompilationScopesToContributions:
-    FirCache<FirSession, Map<ClassId, Set<ClassId>>, TypeResolveService> =
-    session.firCachesFactory.createCache { session, typeResolver ->
-      val scopesToContributingClass = mutableMapOf<ClassId, MutableSet<ClassId>>()
+    FirCache<ClassId, Set<ClassId>, TypeResolveService> =
+    session.firCachesFactory.createCache { scopeClassId, typeResolver ->
       // In a KMP compilation we want to capture _all_ sessions' symbols. For example, if we are
       // generating supertypes for a graph in jvmMain, we want to capture contributions declared in
       // commonMain.
@@ -81,27 +82,23 @@ internal class ContributedInterfaceSupertypeGenerator(session: FirSession) :
           )
         }
         .filterIsInstance<FirRegularClassSymbol>()
-        .forEach { clazz ->
+        .flatMapToSet { clazz ->
           clazz.annotations
             .annotationsIn(session, session.classIds.allContributesAnnotations)
             .mapNotNull { it.resolvedScopeClassId(typeResolver) }
-            .forEach { scopeClassId ->
-              scopesToContributingClass
-                .getOrPut(scopeClassId, ::mutableSetOf)
-                .add(clazz.classId.createNestedClassId(Symbols.Names.metroContribution))
-            }
+            .filter { it == scopeClassId }
+            .map { clazz.classId.createNestedClassId(Symbols.Names.metroContribution) }
         }
-      scopesToContributingClass
     }
 
   private val generatedScopesToContributions:
     FirCache<ClassId, Map<ClassId, Set<ClassId>>, TypeResolveService> =
     session.firCachesFactory.createCache { scopeClassId, typeResolver ->
-      val scopeSimpleName = scopeClassId.joinSimpleNames().shortClassName
+      val scopeHintFqName = Symbols.FqNames.scopeHint(scopeClassId)
       val functionsInPackage =
         session.symbolProvider.getTopLevelFunctionSymbols(
-          Symbols.FqNames.metroHintsPackage,
-          scopeSimpleName,
+          scopeHintFqName.parent(),
+          scopeHintFqName.shortName(),
         )
 
       buildMap<ClassId, MutableSet<ClassId>> {
@@ -187,6 +184,16 @@ internal class ContributedInterfaceSupertypeGenerator(session: FirSession) :
         }
         .filterNotTo(mutableSetOf()) { it == StandardClassIds.Nothing }
 
+    for (classId in scopes) {
+      session.lookupTracker?.recordFqNameLookup(
+        Symbols.FqNames.scopeHint(classId),
+        classLikeDeclaration.source,
+        // The class source is the closest we can get to the file source,
+        // and the file path lookup is cached internally.
+        classLikeDeclaration.source,
+      )
+    }
+
     val contributions =
       scopes
         .flatMap { scopeClassId ->
@@ -196,9 +203,7 @@ internal class ContributedInterfaceSupertypeGenerator(session: FirSession) :
               .orEmpty()
 
           val inCompilationContributions =
-            inCompilationScopesToContributions
-              .getValue(session, typeResolver)[scopeClassId]
-              .orEmpty()
+            inCompilationScopesToContributions.getValue(scopeClassId, typeResolver)
 
           (inCompilationContributions + classPathContributions).map {
             it.constructClassLikeType(emptyArray())